requirements.txt
# === Core AI Inference ===
transformers>=4.38.0
torch>=2.2.0
huggingface_hub>=0.20.0
openai>=1.24.0
replicate>=0.25.0

# === Web Interface ===
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.34.0

# === CLI Framework ===
typer[all]>=0.12.3
python-dotenv>=1.0.1

# === Data Handling ===
pandas>=2.2.0
scikit-learn>=1.4.0

# === Visualization ===
matplotlib>=3.8.0
seaborn>=0.13.0

# === Utilities ===
requests>=2.31.0
tqdm>=4.66.0
aiohttp>=3.9.0
pydantic>=2.6.0
tabulate>=0.9.0

# === Optional Dev Tools ===
ruff>=0.4.0
pytest>=8.2.0
pytest-cov>=5.0.0
